+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - item.data.caspar.data replaces item.data.templateData for structured template data
 ### Added
 - The `item.apply` event
+- A volume item type for the Caspar plugin
 
 ## 1.0.0-beta.5
 ### Changed